requirements.txt
# File: requirements.txt
torch
Pillow
gradio
playwright
loguru
asyncio
python-dotenv
mistune
aiohttp
eth-account
python-slugify
vllm>=0.10.0  # High-throughput inference for multi-step VLM tests (CPU/GPU)
pytest>=8.4.1  # Testing framework for running test suite
bitsandbytes
huggingface_hub